+++ cppwamp/include/cppwamp/internal/udsacceptor.hpp
@@ -44,7 +44,7 @@ public:
             UdsAcceptor* self;
             typename std::decay<F>::type callback;
 
-            void operator()(AsioErrorCode asioEc)
+            void operator()(boost::system::error_code asioEc)
             {
                 SocketPtr socket{std::move(self->socket_)};
                 self->socket_.reset();
@@ -80,7 +80,7 @@ private:
     using Acceptor = boost::asio::local::stream_protocol::acceptor;
 
     template <typename F>
-    bool checkError(AsioErrorCode asioEc, F& callback)
+    bool checkError(boost::system::error_code asioEc, F& callback)
     {
         if (asioEc)
         {